requirements.txt
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
outcome==1.2.0
pkg_resources==0.0.0
PySocks==1.7.1
python-decouple==3.8
requests==2.31.0
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
undetected-chromedriver==3.5.0
urllib3==2.0.3
websockets==11.0.3
wsproto==1.2.0
